VSI 2017 Public release

+++ contrib/brl/bpro/core/vpgl_pro/processes/vpgl_convert_to_local_coordinates_process.cxx
@@ -67,9 +67,9 @@ bool vpgl_convert_to_local_coordinates_process(bprb_func_process& pro)
   double x,y,z;
   lvcs.global_to_local(lon,lat,el, vpgl_lvcs::wgs84, x, y, z, vpgl_lvcs::DEG, vpgl_lvcs::METERS);
 
-  pro.set_output_val<float>(0, (float)x);
-  pro.set_output_val<float>(1, (float)y);
-  pro.set_output_val<float>(2, (float)z);
+  pro.set_output_val<double>(0, (double)x);
+  pro.set_output_val<double>(1, (double)y);
+  pro.set_output_val<double>(2, (double)z);
 
   return true;
 }
@@ -85,20 +85,20 @@ bool vpgl_convert_to_local_coordinates_process2_cons(bprb_func_process& pro)
   bool ok=false;
   std::vector<std::string> input_types;
   input_types.push_back("vpgl_lvcs_sptr");
-  input_types.push_back("float");
-  input_types.push_back("float");
-  input_types.push_back("float");
+  input_types.push_back("double");
+  input_types.push_back("double");
+  input_types.push_back("double");
   ok = pro.set_input_types(input_types);
   if (!ok) return ok;
 
   // outputs x,y,z in local coordinates
-  // 0: (float) x
-  // 1: (float) y
-  // 2: (float) z
+  // 0: (double) x
+  // 1: (double) y
+  // 2: (double) z
   std::vector<std::string> output_types;
-  output_types.push_back("float");
-  output_types.push_back("float");
-  output_types.push_back("float");
+  output_types.push_back("double");
+  output_types.push_back("double");
+  output_types.push_back("double");
   ok = pro.set_output_types(output_types);
   if (!ok) return ok;
 
@@ -116,16 +116,16 @@ bool vpgl_convert_to_local_coordinates_process2(bprb_func_process& pro)
 
   // get the inputs
   vpgl_lvcs_sptr lvcs = pro.get_input<vpgl_lvcs_sptr>(0);
-  float lat = pro.get_input<float>(1);
-  float lon = pro.get_input<float>(2);
-  float el = pro.get_input<float>(3);
+  double lat = pro.get_input<double>(1);
+  double lon = pro.get_input<double>(2);
+  double el = pro.get_input<double>(3);
 
   double x,y,z;
   lvcs->global_to_local(lon,lat,el, vpgl_lvcs::wgs84, x, y, z, vpgl_lvcs::DEG, vpgl_lvcs::METERS);
 
-  pro.set_output_val<float>(0, (float)x);
-  pro.set_output_val<float>(1, (float)y);
-  pro.set_output_val<float>(2, (float)z);
+  pro.set_output_val<double>(0, (double)x);
+  pro.set_output_val<double>(1, (double)y);
+  pro.set_output_val<double>(2, (double)z);
 
   return true;
 }
